#4 - deleteItem 함수 위치 이동에 따른 Test 수정
deleteItem이 worldcupServiceImpl -> ItemServiceImpl 로 이동. 그에 따라 deleteItem을 ItemServiceImplTest로 이동 및 수정했음.
