+++ src/test/java/com/example/ourworldcup/service/worldcup/impl/WorldcupServiceImplTest.java
@@ -47,39 +47,4 @@ class WorldcupServiceImplTest {
                 .hasFieldOrPropertyWithValue("content", content)
                 .hasFieldOrPropertyWithValue("password", password);
     }
-    @DisplayName("월드컵 번호와 아이템 번호가 주어지면, 아이템이 삭제된다.")
-    @Test
-    void givenWorldcupIdAndItemId_whenDeletesItem_thenCorrectlyDeletesItem() {
-        //given
-        Long worldcupId = 1L;
-        Long itemId = 1L;
-        String realUuid = "uuid string";
-        String fileName = "itemImage filename";
-        Worldcup worldcup = Worldcup.builder()
-                .id(worldcupId)
-                .title("worldcup title")
-                .content("worldcup content")
-                .password("worldcup password").build();
-        Uuid uuid = Uuid.builder()
-                .id(1L)
-                .uuid(realUuid)
-                .build();
-        ItemImage itemImage = ItemImage.builder()
-                .id(1L)
-                .url("itemImage url")
-                .fileName(fileName)
-                .uuid(uuid).build();
-
-        Item item = Item.builder()
-                .id(itemId)
-                .worldcup(worldcup)
-                .title("item title")
-                .itemImage(itemImage).build();
-        worldcup.getItems().add(item);
-        given(worldcupRepository.findById(worldcupId)).willReturn(Optional.of(worldcup));
-        //when
-        sut.deleteItem(worldcupId, itemId);
-        //then
-        assertThat(worldcup.getItems()).doesNotContain(item);
-    }
 }
